Guard against not having JS navigation elements
The fsastyles JS code currently assumes that there is a collapsible
menu, whose states will be managed by the JS code in `navigation.js`. In
the case of data catalog, we don't have such a collapsible menu, which
causes errors to be raised from JavaScript.

This commit adds guards to prevent those errors, and changes the
`console.warn` in the case of not detecting JS navigation elements to a
`console.log`, since in our case it is not an error.